Add signature with each key for directory responses
This adds a method to http-message-sig package so that directory
responses can be signed as recommended in the draft. The draft change is
pending and available in https://github.com/thibmeu/http-message-signatures-directory/pull/40

This commit also updates the example worker to provide a signature over
the directory.

Tests are to be added. They could also serve as a basis to get test
vectors in the directory draft.