Fix a confusing type parameter for generating signatures.
The `generate_signature_headers_content` method's `signing_key`
argument references `PublicKey` as its type. That's semantically wrong -
a signing key must be a private key. In this case, `PublicKey` is a type
alias for `Vec<u8>`, which is why it works - but there's no reason for
us to confuse the public. I have removed the `PublicKey` and replaced it
directly with `Vec<u8>`, and bumped Cargo accordingly per semver.