Example updated to use the registry draft
The verification worker now provides a registry on /test-registry.txt.
Caddy plugin has been updated to support fetching registries, and the
signature agent in them.

This is basic, and breaks the old model where caddy could import a
directory directly.

Running the example locally is cumbersome as, beyond chrome extension,
there is no good tool to generate an RFC9421 signed request. That's
something to do later on.

I ran the example by generating a signature with the "scheduled" script.