Simplify interface: concentrate into `lookup_component` method
This change removes the following constructs:

1. `WebBotAuthSignedMessage`
2. `SignedMessage::fetch_all_signature_headers`
3. `SignedMessage::fetch_all_signature_inputs`

and modifies the `lookup_component` method to be the primary source of
information about a message. The goal here is to make it possible to
treat all HTTP headers, including special ones like `Signature-Agent`,
etc. on the same footing. It also removes a source of complexity for
implementors, since they would need to maintain an independent method
just to fetch special values.