Add support for parsing Signature-Agent as an sfv::Dictionary
This change supports handling Signature-Agent as either an
sfv::Dictionary or an sfv::Item. To decide which way to process, it
relies on whether or not the `Signature-Agent` header listed in the
`Signature-Input` field has an associated `key` parameter. If present,
it attempts to parse as a dictionary.

I've added tests to ensure both types can be handled perfectly well.