Support draft-directory-04 with sf-dictionary signature-agent
This commits adds support for sf-dioctionary headers in
http-message-sig, and paired signature-agent as a dictionary format.

This is made to be backward compatible: old test vectors still pass.

The implementation of sf-dictionary is primitive, and likely does not
pass all tests for [RFC 8941](https://www.rfc-editor.org/rfc/rfc8941.html).

This is acceptable for now. We _could_ publish this as an alpha.

The new test vectors are added in
https://github.com/thibmeu/http-message-signatures-directory/pull/79,
and have a corresponding json
[web_bot_auth_architecture_v2.json](./packages/web-bot-auth/test/test_data/web_bot_auth_architecture_v2.json).
They can be imported by other implementations.