Fix registry and http-signature-directory to work with spec
The registry should point to `/well-known/http-message-signature` path, not
just the domain of the host. This change amends that.

`http-signature-directory` has been assuming responses from the
directory will be signed with just the `@authority` component - however, this
is wrong, both per the spec and logically. The correct component should
be `"@authority";req`, indicating that the response is signing the
derived component from the request. This change adds a log line to the
same.

It also adds a better error message for the case when signature
verification fails. This is based on real challenges experienced by the
community e.g. in https://community.cloudflare.com/t/web-bot-auth-signature-validation-failure/854577
during parsing.